+++ lib/src/db/lib_update_dbs.dart
@@ -342,17 +342,18 @@ Future<void> updateIsoneMorningReport(
     for (var day in month.days()) {
       print('Working on $day');
       if (day.isAfter(today)) continue;
-      var file = archive.getFilename(day);
+      var file = archive.getFile(day);
       if (!file.existsSync() || download) {
         var res = await baseDownloadUrl(archive.getUrl(day), file,
             username: dotenv.env['ISONE_WS_USER'],
             password: dotenv.env['ISONE_WS_PASSWORD'],
-            acceptHeader: 'application/json');
+            acceptHeader: 'application/json',
+            gzipFile: true);
         if (res != 0) throw StateError('Failed to download');
       }
     }
-    // archive.makeGzFileForMonth(month);
-    archive.updateDuckDb(month);
+    archive.makeGzFileForMonth(month);
+    archive.updateDuckDB(month);
   }
 }