+++ test/db/isoexpress/morning_report_test.dart
@@ -8,7 +8,7 @@ Future<void> tests() async {
   final archive = getIsoneMorningReportArchive();
   group('ISONE morning report tests:', () {
     test('read file for 2024-01-01', () async {
-      var file = archive.getFilename(Date.utc(2024, 1, 1));
+      var file = archive.getFile(Date.utc(2024, 1, 1));
       var data = archive.processFile(file);
       expect(data.length, 2);
       var columns = data.first.toJson().keys.toList();
@@ -17,21 +17,21 @@ Future<void> tests() async {
     });
     test('read file for 2024-05-31', () async {
       // two new fields for geo magnetic disturbance
-      var file = archive.getFilename(Date.utc(2024, 5, 31));
+      var file = archive.getFile(Date.utc(2024, 5, 31));
       var data = archive.processFile(file);
       expect(data.length, 2);
       var columns = data.first.toJson().keys.toList();
       expect(columns.length, MorningReport.colnames.length);
     });
     test('read file for 2020-09-28', () async {
-      var file = archive.getFilename(Date.utc(2020, 9, 28));
+      var file = archive.getFile(Date.utc(2020, 9, 28));
       var data = archive.processFile(file);
       expect(data.length, 2);
       var columns = data.first.toJson().keys.toList();
       expect(columns.length, MorningReport.colnames.length);
     });
     test('read file for 2023-06-16', () async {
-      var file = archive.getFilename(Date.utc(2023, 6, 16));
+      var file = archive.getFile(Date.utc(2023, 6, 16));
       var data = archive.processFile(file);
       expect(data.length, 2);
       var columns = data.first.toJson().keys.toList();